msm_fb: display: Wait for VSYNC if HDMI connected
Before unsetting and reallcating the pipes, a wait of
one VSYNC is required to avoid mixer getting same pipe
allocated. But in case the HDMI cable is disconnected,
external VSYNC would not be received and the primary
display would not respond. To avoid this, wait during
unset is done only if the HDMI cable is connected.

CRs-fixed: 352950
Change-Id: I416dfbdf5c6d82023781cc8fdbfdb0021898fd18
Signed-off-by: Ajay Singh Parmar <aparmar@codeaurora.org>